fix(sdk): export LuzmoStream for streamed responses
Extract the stream wrapper so consumers can detect streamed responses with instanceof and align the cancellation guards with the review feedback.
